fix(logs): resolve critical bugs and consolidate dual parser architecture
- Remove duplicate command registration in lib.rs that would panic at startup
  (both logs::commands::* and commands::logs::* were registered simultaneously)
- Delete duplicate log_engine module and commands/logs.rs, consolidating to
  the richer logs/ module as the canonical parser
- Fix detect_boss source_health bug: was using target_health instead of
  source_health when checking if a source unit is a boss
- Fix parse_with_offsets CRLF byte offset handling: detect \r\n line endings
  and use correct 2-byte offset for Windows ESO log files
- Move Arc import to top of logs/commands.rs
- Move get_logs_dir into logs/commands.rs, returning Result<String, String>
  instead of silently falling back to empty string
- Fix notify crate to use default features for cross-platform watchers
  (was hardcoded to macos_fsevent only)
- Extract shared formatDuration utility to lib/utils.ts (was duplicated in
  4 files: characters.tsx, encounter-detail.tsx, log-detail.tsx, EncounterList.tsx)
- Remove dead LogsPage and pages/Logs/ directory (LogsWorkspace is the
  active implementation used by App.tsx)
- Fix missing LogsWorkspace import in App.tsx after cleanup
