+++ src/App.tsx
@@ -8,8 +8,8 @@ import { AppBackground } from "./components/app-background";
 import { AppDialogs } from "./components/app-dialogs";
 import { AppHeader } from "./components/app-header";
 import { DiscoverDetail } from "./components/discover-detail";
-import { LogsPage } from "./pages/Logs/LogsPage";
 import { SetupWizard } from "./components/setup-wizard";
+import { LogsWorkspace } from "./components/logs/logs-workspace";
 import { StatusBanners } from "./components/status-banners";
 import { UpdateBanner } from "./components/update-banner";
 import { getSetting, setSetting } from "@/lib/store";
