+++ src/components/logs/log-detail.tsx
@@ -5,6 +5,7 @@ import { GlassPanel } from "@/components/ui/glass-panel";
 import { InfoPill } from "@/components/ui/info-pill";
 import { SectionHeader } from "@/components/ui/section-header";
 import type { LogAnalysis, Encounter } from "@/types/logs";
+import { formatDuration } from "@/lib/utils";
 
 interface LogDetailProps {
   analysis: LogAnalysis;
@@ -153,8 +154,3 @@ export function LogDetail({ analysis, onBack, onOpenEncounter }: LogDetailProps)
   );
 }
 
-function formatDuration(secs: number): string {
-  const m = Math.floor(secs / 60);
-  const s = Math.floor(secs % 60);
-  return `${m}:${s.toString().padStart(2, "0")}`;
-}
