+++ src/lib/utils.ts
@@ -12,3 +12,9 @@ export function decodeHtml(str: string): string {
   _htmlDecodeEl.innerHTML = str;
   return _htmlDecodeEl.value;
 }
+
+export function formatDuration(secs: number): string {
+  const m = Math.floor(secs / 60);
+  const s = Math.floor(secs % 60);
+  return `${m}:${s.toString().padStart(2, "0")}`;
+}
